+++ run_tests.sh
@@ -11,6 +11,7 @@ NUMBER_2="$2"
 TEST_PIN_1=456test_pin_foo123
 NATIVE=0
 JSON_RPC=0
+DBUS=0
 TEST_REGISTER=0
 
 PATH_TEST_CONFIG="$PWD/test-config"
@@ -41,8 +42,10 @@ run() {
   set -x
   if [ "$JSON_RPC" -eq 1 ]; then
     "$SIGNAL_CLI" $@
+  elif [ "$DBUS" -eq 1 ]; then
+    "$SIGNAL_CLI" --dbus --verbose --verbose $@ | grep -v '^Warning:' | grep -v 'at org'
   else
-    "$SIGNAL_CLI" --service-environment="staging" --verbose --verbose $@
+    "$SIGNAL_CLI" --service-environment="staging" --verbose --verbose $@ | grep -v '^Warning:' | grep -v 'at org'
   fi
   set +x
 }
@@ -261,6 +264,6 @@ fi
 run_main -a "$NUMBER_2" deleteLocalAccountData || true
 
 if [ ! -z "$GRAALVM_HOME" ]; then
-  "$GRAALVM_HOME"/lib/svm/bin/native-image-configure generate --input-dir=graalvm-config-dir/ --input-dir=graalvm-config-dir-linked/ --input-dir=graalvm-config-dir-main/ --output-dir=graalvm-config-dir//
+  "$GRAALVM_HOME"/lib/svm/bin/native-image-configure generate --input-dir=src/main/resources/META-INF/native-image/org.asamk/signal-cli/ --input-dir=graalvm-config-dir-linked/ --input-dir=graalvm-config-dir-main/ --output-dir=src/main/resources/META-INF/native-image/org.asamk/signal-cli/
   rm -r graalvm-config-dir-main graalvm-config-dir-linked
 fi